crop pad return params

+++ batchgenerators/augmentations/crop_and_pad_augmentations.py
@@ -59,6 +59,7 @@ def get_lbs_for_constrained_random_crop(crop_size, data_shape, anchor, margins, 
     :param rs:
     :return:
     """
+    anchor = anchor.astype(np.int)
     lbs = []
     for i in range(len(data_shape) - 2):
         margin_left = anchor[i] - margins[i]
@@ -68,7 +69,7 @@ def get_lbs_for_constrained_random_crop(crop_size, data_shape, anchor, margins, 
         if margin_right > data_shape[i+2]-crop_size[i]//2:
             margin_right = data_shape[i+2]-crop_size[i]//2
         if margin_left-crop_size[i]//2 >= margin_right-crop_size[i]//2:
-            lbs.append(int(anchor[i] - crop_size[i]//2))
+            lbs.append(anchor[i] - crop_size[i]//2)
         else:
             if rs is None:
                 a = np.random.randint(margin_left-crop_size[i]//2, margin_right-crop_size[i]//2 + 1)
